dashboards: use custom all value for instance selector
Currently, instance selector is only populated on dashboard load. This leads to panels being empty when extending time
range if instances list was updated, this is frequently happening in dynamic environments like k8s.

Fix this by using custom `.*` value so that all instances are always matched without need to re-populate the list of
instances.